requirements.txt
pandas
requests
python-dotenv
pyarrow
matplotlib
seaborn
jupyterlab
ipykernel
jupyter
notebook
